Move TreeBuilder to the correct namespace
